Groundbreaking new features
This commit introduces a detailed multi-layered scrolling ground. The new
ground scrolls in parallax with the pipes and is composed of several distinct
visual layers. It consist of a top grass layer with a looping chevron pattern,
while the larger dirt section below it uses basic procedural generation to
create a non-repeating texture of mud specks and rocks. Also the game has been
updated to no longer reset instantly upon collision/death.